cleanup of index legacy and moved scenemanager
