--- Assets/Scripts/index.html
@@ -1,62 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-
-<head>
-  <meta charset=utf-8>
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta name="author" content="Gerard Llorach">
-
-</head>
-<body>
-
-<script type="module">
-
-  import WMTSDataRetriever from "./WMTSDataRetriever.js"
-
-  let getWMTSData = () => {
-
-    wmts.dataSets.forEach(dataSet => {
-      let container = document.createElement('div');
-      container.style.display = 'flex';
-      container.style["flex-direction"] = 'row';
-
-      let text = document.createElement('span');
-      let tmst = new Date().toISOString();// '2022-05-31T00:00:00Z'; // (dataSet.endTmst || 'now') new Date().toISOString()
-      text.innerHTML = dataSet.id + ", " + dataSet.timeScale +  ", " +  tmst;
-      //let img = wmts.getImageAt(dataSet, new Date().toISOString());
-      let img = wmts.getDataAtPoint(dataSet.id, tmst, 41.3, 2.2, dataSet.timeScale).then(value => {
-        if (value != undefined){
-          try{
-            console.log(dataSet.name + " " + value.toFixed(2) + " " + dataSet.unit + " at " + dataSet.timeScale);
-          }
-          catch{
-            debugger;
-          }
-          
-          }
-      });
-
-      container.append(img);
-      container.append(text);
-      document.body.append(container);
-    });
-
-    // let dtset = wmts.getDataSet('VHM0', 'h', new Date().toISOString());
-
-    // let img = wmts.getImageAt(dtset, new Date().toISOString());
-
-    // document.body.append(img);
-
-    // console.log(dtset);
-  }
-
-  let wmts = new WMTSDataRetriever(getWMTSData, true);
-
- 
-
-</script>
-
-
-
-</body>
-</html>
